refactor: Move to TS && use latest dependencies && support latest textlint
