+++ package.json
@@ -20,22 +20,26 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/azu/textlint-rule-first-sentence-length.git"
+    "url": "https://github.com/textlint-rule/textlint-rule-first-sentence-length.git"
   },
   "bugs": {
-    "url": "https://github.com/azu/textlint-rule-first-sentence-length/issues"
+    "url": "https://github.com/textlint-rule/textlint-rule-first-sentence-length/issues"
   },
-  "homepage": "https://github.com/azu/textlint-rule-first-sentence-length",
+  "homepage": "https://github.com/textlint-rule/textlint-rule-first-sentence-length",
   "devDependencies": {
-    "textlint-scripts": "^1.2.2",
-    "textlint-tester": "^2.1.0"
+    "@textlint/ast-node-types": "^13.3.0",
+    "@textlint/types": "^13.3.0",
+    "textlint-scripts": "^13.3.0",
+    "textlint-tester": "^13.3.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.5"
   },
   "keywords": [
     "textlintrule"
   ],
   "dependencies": {
-    "sentence-splitter": "^2.1.1",
+    "sentence-splitter": "^4.0.2",
     "textlint-rule-helper": "^2.0.0",
-    "textlint-util-to-string": "^2.0.1"
+    "textlint-util-to-string": "^3.3.0"
   }
 }
